+++ packages/dapp/components/Background/Background.js
@@ -1,12 +1,8 @@
-import Image from 'next/image';
-
-import BackgroundImage from '../../public/assets/dark-background.png';
-
 import styles from './Background.module.scss';
 
 const Background = () => (
   <div className={styles.bgWrap}>
-    <Image src={BackgroundImage} layout="fill" objectFit="cover" />
+    <img src={'/assets/dark-background.png'} />
   </div>
 );
