+++ packages/dapp/components/Hero/Hero.js
@@ -1,17 +1,13 @@
 import { useRouter } from 'next/router';
-import Image from 'next/image';
-
-import HomeTitle from '../../public/assets/ethernaut-home-title.svg';
-import GalleryTitle from '../../public/assets/ethernaut-gallery-title.svg';
 
 import styles from './Hero.module.scss';
 
 import { MAIN_ROUTES, GALLERY_ROUTES } from '../../constants/routes';
 
 const heroKind = {
-  [MAIN_ROUTES.home.path]: <Image src={HomeTitle} />,
-  [GALLERY_ROUTES.all.path]: <Image src={GalleryTitle} />,
-  [GALLERY_ROUTES.me.path]: <Image src={GalleryTitle} />,
+  [MAIN_ROUTES.home.path]: <img src={'/assets/ethernaut-home-title.svg'} />,
+  [GALLERY_ROUTES.all.path]: <img src={'/assets/ethernaut-gallery-title.svg'} />,
+  [GALLERY_ROUTES.me.path]: <img src={'/assets/ethernaut-gallery-title.svg'} />,
 };
 
 const Hero = () => {
