+++ packages/dapp/components/Logs/styles.module.scss
@@ -37,7 +37,7 @@
     height: 20%;
     display: flex;
     justify-content: space-between;
-    align-items: end;
+    align-items: flex-end;
 
     .button {
       background: none;
