+++ packages/dapp/config.js
@@ -3,6 +3,7 @@ const isDev = process.env.NODE_ENV === 'development';
 let TOKEN_ADDRESS = process.env.NEXT_PUBLIC_TOKEN_ADDRESS;
 const ETHEREUM_NETWORK = process.env.NEXT_PUBLIC_ETHEREUM_NETWORK;
 const INFURA_PROJECT_ID = process.env.NEXT_PUBLIC_INFURA_PROJECT_ID;
+const FLEEK_BUCKET_ID = process.env.NEXT_PUBLIC_FLEEK_BUCKET_ID;
 
 if (!TOKEN_ADDRESS && isDev) {
   (async () => {
@@ -17,7 +18,6 @@ if (isDev) {
   ABI = require('@ethernauts/hardhat/artifacts/contracts/Ethernauts.sol/Ethernauts.json').abi;
 } else {
   ABI = JSON.parse(process.env.NEXT_PUBLIC_CONTRACT_ABI);
-  console.log('ABI: ', ABI);
 }
 
-export { TOKEN_ADDRESS, ETHEREUM_NETWORK, INFURA_PROJECT_ID, ABI };
+export { TOKEN_ADDRESS, ETHEREUM_NETWORK, INFURA_PROJECT_ID, FLEEK_BUCKET_ID, ABI, isDev };
