+++ packages/dapp/next.config.js
@@ -8,6 +8,9 @@ module.exports = {
       '/gallery/me': { page: '/gallery/me' },
     };
   },
+  images: {
+    domains: ['https://storageapi.fleek.co/'],
+  },
   // Fix fast refresh for Windows users
   webpackDevMiddleware: (config) => {
     config.watchOptions = {
