+++ packages/dapp/public/assets/revealed-token.png
