+++ packages/dapp/public/assets/unrevealed-token.png
