+++ packages/dapp/styles/react-medium-image-zoom.scss
@@ -1,4 +1,5 @@
 [data-rmiz-wrap="visible"],
 [data-rmiz-wrap="hidden"] {
   height: inherit;
+  width: inherit;
 }
